docs: translate comments and docs to English
